requirements.txt
python-telegram-bot==20.7
yfinance==0.2.36
pytz==2024.1
